make sure neuron blacklist contains ints

+++ tools/synapse_driver_check/defects.py
@@ -293,7 +293,7 @@ if __name__ == "__main__":
 
     random.seed(1)
 
-    neuron_blacklist = np.loadtxt('blacklist_w{}_h{}.csv'.format(WAFER, HICANN)) # no notion of freq and bkgisi
+    neuron_blacklist = np.loadtxt('blacklist_w{}_h{}.csv'.format(WAFER, HICANN), dtype=int) # no notion of freq and bkgisi
 
     hardware = shallow.Hardware(WAFER, HICANN, os.path.join(args.calibpath,'wafer_{0}').format(WAFER), FREQ*1e6, BKGISI)
